Fixed defineProperty error
There is an error - TypeError: Attempting to change enumerable attribute
of unconfigurable property.

Function's name is not configurable in old browsers (IE or Safari).
`Note that in non-standard, pre-ES2015 implementations the configurable
attribute was false as well.`
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/name
I decided that name prop is not so important and we can neglect it